Rename ws package to rpc
