+++ rpc/client.go
@@ -1,4 +1,4 @@
-package ws
+package rpc
 
 import (
 	"github.com/0xProject/0x-mesh/zeroex"
